requirements.txt
# 核心OCR引擎
paddlepaddle>=2.5.0
paddleocr>=2.7.0
easyocr>=1.7.0

# 图像处理
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0

# Web API
flask>=3.0.0
flask-cors>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# 配置管理
PyYAML>=6.0
python-dotenv>=1.0.0

# 命令行工具
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0

# 工具库
requests>=2.31.0
validators>=0.22.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 日志
loguru>=0.7.0
